Move COM basics to Core & convert more GDI+ interop
This change converts the Image and Bitmap classes to CsWin32.

- Move IStream, ComHelpers, IDispatch, VARIANT and other core COM code to Core
- Change the name of GpStream to ComManagedStream and consolidate both versions
- Create an extension method for Stream to create an IStream scope
- Take in the latest CsWin32 drop to fix a blocking COM generation issue
- Use the new option to not generate friendly overloads in System.Drawing (there is a pending metadata issue that causes some of these to be generated unsafely).
- Update some code to throw ArgumentNull instead of NullRef
- Fix PrinterSettings.InstalledPrinters and add a test

+++ src/System.Drawing.Common/src/System/Drawing/Imaging/CachedBitmap.cs
@@ -44,7 +44,7 @@ public unsafe sealed class CachedBitmap : IDisposable
         ArgumentNullException.ThrowIfNull(graphics);
 
         Gdip.CheckStatus(Gdip.GdipCreateCachedBitmap(
-            new(bitmap, bitmap._nativeImage),
+            new(bitmap, (nint)bitmap._nativeImage),
             new(graphics, (nint)graphics.NativeGraphics),
             out _handle));
     }
